frontend/requirements.txt
streamlit==1.31.1
qdrant-client==1.7.0
sentence-transformers==2.5.1
psycopg2-binary==2.9.9
pandas==2.2.0
requests==2.31.0
python-dotenv==1.0.1